changed `calc.py` to specify `np.int64` data types

+++ src/calc.py
@@ -117,12 +117,12 @@ def _roll_k_iter(roll_1: SequenceWithOffset):
 
 def roll_1dn(n: int):
     if n == 0:
-        return SequenceWithOffset(seq=np.zeros(0), offset=0)
-    return SequenceWithOffset(seq=np.full(n, fill_value=1), offset=1)
+        return SequenceWithOffset(seq=np.zeros(0, dtype=np.int64), offset=0)
+    return SequenceWithOffset(seq=np.full(n, fill_value=1, dtype=np.int64), offset=1)
 
 
 def roll_0():
-    return SequenceWithOffset(seq=np.ones(1), offset=0)
+    return SequenceWithOffset(seq=np.ones(1, dtype=np.int64), offset=0)
 
 
 def _take_index_n(iter, n: int):
